More Sanitation and error handling
Added a check before creating the user to see if the email already exists in the Workspace account. If it does, suggest a new email address and allow user to accept the new address. If the email address does not exist, then we are free to create a user with that address.

Added a check for the Employee's name fields. Check for spaces and confirm that they should exist, make sure to strip hyphens and spaces from the name while creating the username.